Implement SAM_TIMEOUT_FAULT and use it in a test

+++ src/main/java/bio/terra/service/dataset/flight/ingest/IngestCleanupStep.java
@@ -39,8 +39,10 @@ public class IngestCleanupStep implements Step {
         try {
             Dataset dataset = IngestUtils.getDataset(context, datasetService);
 
+            // overlappingTableName is null when the ingest strategy is not upsert.
+            // Don't try to delete the table name if it is null
             overlappingTableName = IngestUtils.getOverlappingTableName(context);
-            if (bigQueryPdao.tableExists(dataset, overlappingTableName)) {
+            if (overlappingTableName != null && bigQueryPdao.tableExists(dataset, overlappingTableName)) {
                 bigQueryPdao.deleteDatasetTable(dataset, overlappingTableName);
             }
         } catch (Exception ex) {
